+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 
+## 1.2.0
+- Add `cigar.to_char_pairs(pattern, text)` to iterate corresponding characters.
+
 ## 1.1.0
-- #2: CIGARs also include 1.
+- #2: `cigar.to_string()` explicitly writes counts of 1.
 
 ## 1.0.0
 - Initial release